# Specialize empty key value pairs