Currently, the `__private_api::log` function takes `None::<&[(&str, _)]>` as the `kvs` argument if no key-value pair is provided, but passing a `None` value may have some extra cost. We should be able to avoid this cost by encoding this information into a zero sized type (here I am using `()`, not sure if it is better to use a dedicated new type (such as `struct EmptyKVs;`)). Constructing a zero sized value is zero cost, which is less than constructing a `None::<&[(&str, _)]>`.